35 gekko analyzer

+++ SRC/Core/Gekko.h
@@ -7,6 +7,7 @@
 #include <list>
 #include <atomic>
 #include "GekkoDefs.h"
+#include "GekkoAnalyzer.h"
 
 // TODO: Get rid of this non-incapsulated mess
 
@@ -146,7 +147,7 @@ typedef struct CPUControl
 
     int64_t     one_second;         // one second in timer ticks
     bool        decreq;             // decrementer exception request
-    uint32_t    ops;                // instruction counter (only for debug!)
+    size_t      ops;                // instruction counter (only for debug!)
 
     // for default interpreter
     bool        exception;          // exception pending
